requirements-test.txt
decorator  # dependency of swh.core[http]
djangorestframework-stubs
django-stubs
hypothesis
pytest
pytest-django
pytest-mock
requests-mock
swh.core[http] >= 0.0.95
swh.loader.git >= 0.1.2
swh.storage >= 0.1.1
